requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
oauthlib==3.2.2
praw==7.7.0
prawcore==2.3.0
python-slugify==8.0.1
requests==2.31.0
requests-oauthlib==1.3.1
text-unidecode==1.3
update-checker==0.18.0
urllib3==2.0.3
websocket-client==1.6.0
